registry/requirements.txt
flask >= 2.0.1
flask-cors >= 4.0.0
opentelemetry-distro==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc>=1.17.0